refactor(ui): extract creation steps and shared selectors
Move character creation state machine to _creation_steps.py and
subclass/race/class pickers to _selectors.py; replace asserts with guards.
